Predicates are relocated
