+++ core/predicates.go
@@ -1,6 +1,9 @@
 package core
 
-import . "github.com/hazelcast/go-client/serialization"
+import (
+	. "github.com/hazelcast/go-client/internal/predicates"
+	. "github.com/hazelcast/go-client/serialization"
+)
 
 func Sql(sql string) IPredicate {
 	return NewSqlPredicate(sql)
