+++ internal/predicates/predicates_impl.go
@@ -1,4 +1,6 @@
-package serialization
+package predicates
+
+import . "github.com/hazelcast/go-client/serialization"
 
 const PREDICATE_FACTORY_ID = -32
